Fix issue with Dirfile samples being dropped when writing to NFS mounts.
Without this change, CLASS was experiencing dropped Dirfile samples when
writing to NFS mounts; there were no issues when writing to a local disk.

I was never able to track down the exact cause of the issue, and I don't know
if the fix is optimal. However, it successfully fixed the problem CLASS was
having.